requirements.txt
anthropic
mcp-core
langgraph